+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.menugroup.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.menugroup.application.MenuGroupService;
-import kitchenpos.menugroup.ui.dto.MenuGroupCreateRequest;
-import kitchenpos.menugroup.ui.dto.MenuGroupResponse;
+import kitchenpos.menugroup.application.dto.MenuGroupCreateRequest;
+import kitchenpos.menugroup.application.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
