+++ src/main/java/kitchenpos/order/application/dto/ChangeOrderStatusRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.ui.dto;
+package kitchenpos.order.application.dto;
 
 import kitchenpos.order.domain.OrderStatus;
