+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.product.ui;
 
 import kitchenpos.product.application.ProductService;
-import kitchenpos.product.ui.dto.ProductCreateRequest;
-import kitchenpos.product.ui.dto.ProductResponse;
+import kitchenpos.product.application.dto.ProductCreateRequest;
+import kitchenpos.product.application.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
