+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.table.ui.dto.ChangeEmptyRequest;
-import kitchenpos.table.ui.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.table.ui.dto.OrderTableCreateRequest;
-import kitchenpos.table.ui.dto.OrderTableResponse;
+import kitchenpos.table.application.dto.ChangeEmptyRequest;
+import kitchenpos.table.application.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.table.application.dto.OrderTableCreateRequest;
+import kitchenpos.table.application.dto.OrderTableResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
