+++ src/main/java/kitchenpos/table/application/dto/ChangeNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.ui.dto;
+package kitchenpos.table.application.dto;
 
 public class ChangeNumberOfGuestsRequest {
