+++ src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -3,10 +3,10 @@ package kitchenpos.table.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.table.application.TableService;
-import kitchenpos.table.ui.dto.ChangeEmptyRequest;
-import kitchenpos.table.ui.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.table.ui.dto.OrderTableCreateRequest;
-import kitchenpos.table.ui.dto.OrderTableResponse;
+import kitchenpos.table.application.dto.ChangeEmptyRequest;
+import kitchenpos.table.application.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.table.application.dto.OrderTableCreateRequest;
+import kitchenpos.table.application.dto.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
