+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -18,10 +18,10 @@ import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.ui.dto.ChangeOrderStatusRequest;
-import kitchenpos.order.ui.dto.OrderCreateRequest;
-import kitchenpos.order.ui.dto.OrderLineItemRequest;
-import kitchenpos.order.ui.dto.OrderResponse;
+import kitchenpos.order.application.dto.ChangeOrderStatusRequest;
+import kitchenpos.order.application.dto.OrderCreateRequest;
+import kitchenpos.order.application.dto.OrderLineItemRequest;
+import kitchenpos.order.application.dto.OrderResponse;
 import kitchenpos.product.domain.Product;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableStatus;
