+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -18,9 +18,9 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
 import kitchenpos.product.domain.Product;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.ui.dto.OrderTableResponse;
-import kitchenpos.table.ui.dto.TableGroupCreateRequest;
-import kitchenpos.table.ui.dto.TableGroupResponse;
+import kitchenpos.table.application.dto.OrderTableResponse;
+import kitchenpos.table.application.dto.TableGroupCreateRequest;
+import kitchenpos.table.application.dto.TableGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
